+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Master
 
 - Support for parsing lazy vars into Variable's attributes.
+- Updated Stencil to 0.13.1 and SwiftStencilKit to 2.7.0
 - In Swift templates CLI arguments should now be accessed via `argument` instead of `arguments`, to be consistent with Stencil and JS templates.
 
 ## 0.15.0